requirements.txt
pandas>=1.5
requests>=2.28
python-dotenv>=1.0
matplotlib>=3.6
reportlab>=3.6
tabulate>=0.9
tqdm>=4.65
